create reusable animatebars mixin
